slight rework of layout
Changed the look of a few action links and the positioning of some information.
Added link to index in the navbar and admin features
attempted to fix "Recent Post" sections left in reworked code as it presents the same error. Will revisit and remove feature if a solution is not found.